add router and page kirill
