+++ src/components/header/header.css
@@ -1,0 +1,12 @@
+header nav ul {
+    display: flex;
+}
+
+header nav li {
+    list-style: none;
+    margin: 0 20px;
+}
+
+header nav a {
+    list-style: none;
+}
