+++ src/components/header/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Header';
