+++ src/pages/alex/Alex.jsx
@@ -1,0 +1,57 @@
+import { useState } from 'react';
+
+
+const send = (a, b , operator) => {
+  const body = JSON.stringify({ a, b});
+  const headers = { 'Content-Type': 'application/json'};
+  return fetch(`http://localhost:3001/calc/${operator}`, {method:'POST',headers, body})
+}
+
+const App = () => {
+  const [a, setA] = useState(0);
+  const [b, setB] = useState(0);
+  const [result, setResult] = useState([]);
+  const [operator, setOperator] = useState('+');
+
+  const onAChange = (e) => setA(+e.target.value || 0);
+  const onBChange = (e) => setB(+e.target.value || 0);
+  const onOperatorChange = (op) => {
+    setOperator(op);
+  };
+  const sumbite = () => {
+    send(a, b, operator).then(res => res.json()).then(res => {
+      setResult(res.result)
+    })
+  };
+  return (
+    <div className="App">
+      <div>
+        <div className="values">
+          <input type='number' onChange={onAChange} />
+          <input type='number' onChange={onBChange} />
+        </div>
+        <div>
+          <Operator selected={operator} onClick={onOperatorChange} operator='plus' />
+          <Operator selected={operator} onClick={onOperatorChange} operator='minus' />
+          <Operator selected={operator} onClick={onOperatorChange} operator='multiply' />
+          <Operator selected={operator} onClick={onOperatorChange} operator='divide' />
+          <Operator selected={operator} onClick={onOperatorChange} operator='involve' />
+          <Operator selected={operator} onClick={onOperatorChange} operator='remainder' />
+          <Operator onClick={sumbite} operator='=' />
+        </div>
+      </div>
+      <p>Result: {result}</p>
+      <hr />
+    </div>
+  );
+}
+
+const Operator = ({ operator, onClick, selected }) => {
+  return (
+    <button className='operator-btn' disabled={selected === operator} onClick={() => onClick(operator)}>{operator}</button>
+  )
+}
+
+
+
+export default App;
